Fix Goodreads import crash on decimal ratings
- My Rating values like "3.0" broke int() parsing; score is a
  DecimalField(decimal_places=1) so float() works fine
- Add regression test for decimal ratings

Fixes #379

+++ src/integrations/imports/goodreads.py
@@ -249,7 +249,7 @@ class GoodReadsImporter:
         instance = model(
             item=item,
             user=self.user,
-            score=None if rating in {"", "0"} else int(rating) * 2,
+            score=None if rating in {"", "0"} else float(rating) * 2,
             progress=book_progress,
             status=book_status,
             end_date=date_rated,
